+++ src/main/java/innowise/internship/onlineshop/services/OrderService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface OrderService {
     void save(OrderDto order);
     List<OrderDto> getAll();
-    OrderDto getById(int id);
+    OrderDto getById(Long id);
     void update(OrderDto order);
-    void delete(int id);
+    void delete(Long id);
 }
